requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
redis==5.0.3
kubernetes==29.0.0
langchain==0.1.13
boto3==1.34.0
langchain-aws==0.1.0
pydantic-settings==2.2.1
